Refactoring pokemon lib into lib/pokemon and lib/pokemonDetail
